+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/client/PrisonApiClient.kt
@@ -24,9 +24,6 @@ class PrisonApiClient(
       .retrieve()
       .bodyToMono(offenderNonAssociationDetails)
       .block(apiTimeout)
-      .also {
-        log.debug("Get offender non-association details for $offenderNo")
-      }
   }
 
   companion object {
