+++ src/main/kotlin/uk/gov/justice/digital/hmpps/visitscheduler/service/SupportService.kt
@@ -1,0 +1,15 @@
+package uk.gov.justice.digital.hmpps.visitscheduler.service
+
+import org.springframework.stereotype.Service
+import uk.gov.justice.digital.hmpps.visitscheduler.data.SupportTypeDto
+import uk.gov.justice.digital.hmpps.visitscheduler.jpa.repository.SupportTypeRepository
+
+@Service
+class SupportService(
+  private val supportTypeRepository: SupportTypeRepository,
+) {
+
+  fun getSupportTypes(): List<SupportTypeDto> {
+    return supportTypeRepository.findAll().sortedBy { it.code }.map { SupportTypeDto(it) }
+  }
+}
