--- src/main/resources/db/migration/V1_4__create_session_exception_table.sql
@@ -1,12 +1,0 @@
-CREATE TABLE session_exception
-(
-    id                      serial          NOT NULL PRIMARY KEY,
-    session_template_id     serial,
-    prison_id               VARCHAR(6)      NOT NULL,
-    open_capacity           integer         NOT NULL,
-    closed_capacity         integer         NOT NULL,
-    end_date                date,
-    start_date              date            NOT NULL,
-    expiry_date             date,
-    reason                  text
-);
