+++ src/main/resources/db/migration/V1_6__create_visit_notes_table.sql
